Add reasonable application command handling architecture, and comments. (#19)
* Refactor Rust bot to handle commands of different types.

This is a very rough high-level outline of application command
handling.

* Add reasonable application command handling architecture, and comments.

This commit also adds a rustfmt.toml file (though it contains options
which only work on the nightly rustfmt toolchain). We could simplify
the list of format options by a fair bit, but a few very useful options
do still require nightly (`wrap_comments`, for instance).

This commit also adds documentation for all implemented features.
Documentation might be a bit rough around the edges, but it should
get the point across, subject to some proof-reading.